pagination fixies

+++ server/Services/EventsService.cs
@@ -64,8 +64,8 @@ namespace Warehouse2.Services
                 }
             }
 
-            decimal d = allEvents.Count / 7.0;
-            page.count = Math.Ceiling(d);
+            double d = allEvents.Count / 7.0f;
+            page.count = ((int)Math.Ceiling(d));
 
             return page;
         }
@@ -97,8 +97,8 @@ namespace Warehouse2.Services
                 }
             }
 
-            decimal d = cells.Count / 7.0;
-            rentedCells.count = Math.Ceiling(d);
+            double d = cells.Count / 7.0f;
+            rentedCells.count = ((int)Math.Ceiling(d));
 
             return rentedCells;
         }
